nochungus2020
-added the no chungus challenge

Goal: Go as long as you can through 2020 without saying "chungus" or any of its derivatives (chungi, chunglord, etc.)
Times will be tracked on a shoobopper webpage (this post will be edited when the challenge is effective.

Notes:
- The challenge relies on honesty. If you have said chungus in private, come clean and shoot me at shoobopper#2475.
- The challenge uses the eastern time zone. Times will be recorded with this in mind.
- Saying chungus is still allowed in image and text posts.
- You are able to enter the chungus challenge after the new year, as long as you have not said it since the new year.

Reply or react to this post to be entered into the chungus challenge.